Make `GlobalExecutor` part of `concurrent-api` module
Motivation:

`Executors.immediate()` was used in time sensitive operators, which
don't accept an `Executor` instance. When the `ImmediateExecutor` was
used to offload a scheduled task, such as a timeout, it would have been
executed on the global
`io.servicetalk.concurrent.api.DefaultExecutor.SingleThreadedScheduler`
which is used for all future executions internally. When the actual
termination happened, if it blocked, it would have stalled all following
scheduled delayed tasks.

Modifications:

- Moved `GlobalExecutor` from `transport-netty-internal` module to
  `concurrent-api`,
- Introduced `Executors.global()` `Executor`, which uses a singleton of
  `GlobalExecutor`,
- Using `Executors.global()` in all `timeout(...)` operators that do not
  accept `Executor` as an argument.

Result:

The offloading of terminating a `Timeout*` operators happens on the
`GlobalExecutor`, which is the default concurrent instance also used for
reactive chain processing in all transport related operations.
